Added FadeController to UI, TMPro
Imported TMPro for text as Unity's is trash. Set up a UI canvas in the 'Dungeon' scene with a text panel that removes itself on play. Pressing 'i' will fade in the text panel and pressing 'o' will fade out the text panel. This is for testing purposes but is proof of concept.